ansible/requirements.txt
yandexcloud>=0.109.0
pyyaml>=6.0
